refactor(repository): UserRepository를 인터페이스로 만들고 JCFUserRepository 클래스를 생성
- 원래 작성했던 UserRepository클래스를 JCFUserRepository로 옮김
